2.历史版本/ATimer-0.0.3/requirements.txt


click==8.0.4
Flask==2.0.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.1.1
Werkzeug==2.0.3

pyyaml==6.0
markdown==3.4.1
protobuf==3.20.1

matplotlib==3.6.2
seaborn==0.12.1
numpy==1.23.5
pandas==1.5.2

pytest==7.2.0